+++ CHANGELOG.md
@@ -29,6 +29,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `jj debug config-schema` command prints out JSON schema for the jj TOML config
   file format.
 
+* `jj resolve` now notifies the user of remaining conflicts, if any, on success.
+  This can be prevented by the new `--quiet` option.
+
 ### Fixed bugs
 
 * When sharing the working copy with a Git repo, we used to forget to export